requirements.txt
flask
requests
python-dotenv
flask_sqlalchemy